requirements.txt
GitPython==2.1.5
PrettyTable==0.7.2
requests==2.19.1
urllib3==1.23